+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- CIP68 `getMetadataFromOutputDatum` parsing of `files` and custom fields
+
 ## [2.3.0] - 2023-02-02
 
 ### Added